Do not include private header from public header
